Allow tool selection during assistant setup or subsequent assistant update
