+++ routers/setup.py
@@ -1,13 +1,14 @@
 import logging
 import os
-from typing import Optional
+from typing import Optional, List
 from dotenv import load_dotenv
 from fastapi import APIRouter, Depends, HTTPException, Form, Request
 from fastapi.responses import RedirectResponse, Response
 from fastapi.templating import Jinja2Templates
 from openai import AsyncOpenAI
+from openai.types.beta import Assistant
 
-from utils.create_assistant import create_or_update_assistant, request as assistant_create_request
+from utils.create_assistant import create_or_update_assistant, ToolTypes
 from utils.create_assistant import update_env_file
 
 # Configure logger
@@ -45,27 +46,44 @@ async def set_openai_api_key(api_key: str = Form(...)) -> RedirectResponse:
 
 # Add new setup route
 @router.get("/")
-async def read_setup(request: Request, message: Optional[str] = "") -> Response:
+async def read_setup(
+    request: Request,
+    client: AsyncOpenAI = Depends(lambda: AsyncOpenAI()),
+    message: Optional[str] = ""
+) -> Response:
     # Check if assistant ID is missing
+    current_tools = []
     load_dotenv(override=True)
     openai_api_key = os.getenv("OPENAI_API_KEY")
     assistant_id = os.getenv("ASSISTANT_ID")
     
     if not openai_api_key:
         message = "OpenAI API key is missing."
-    elif not assistant_id:
-        message = "Assistant ID is missing."
     else:
-        message = "All set up!"
+        if assistant_id:
+            try:
+                assistant = await client.beta.assistants.retrieve(assistant_id)
+                current_tools = [tool.type for tool in assistant.tools]
+            except Exception as e:
+                logger.error(f"Failed to retrieve assistant {assistant_id}: {e}")
+                # If we can't retrieve the assistant, proceed as if it doesn't exist
+                assistant_id = None
+                message = "Error retrieving existing assistant. Please create a new one."
     
     return templates.TemplateResponse(
         "setup.html",
-        {"request": request, "message": message}
+        {
+            "request": request,
+            "message": message,
+            "assistant_id": assistant_id,
+            "current_tools": current_tools
+        }
     )
 
 
 @router.post("/assistant")
 async def create_update_assistant(
+    tool_types: List[ToolTypes] = Form(...),
     client: AsyncOpenAI = Depends(lambda: AsyncOpenAI())
 ) -> RedirectResponse:
     """
@@ -76,7 +94,7 @@ async def create_update_assistant(
     new_assistant_id = await create_or_update_assistant(
         client=client,
         assistant_id=current_assistant_id,
-        request=assistant_create_request,
+        tool_types=tool_types,
         logger=logger
     )
