+++ static/styles.css
@@ -19,6 +19,7 @@ body {
 }
 
 body {
+  font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
   color: rgb(var(--foreground-rgb));
 }
 
@@ -57,16 +58,8 @@ pre {
 .logo {
   width: 32px;
   height: 32px;
-  position: absolute;
-  margin: 16px;
-  top: 0;
-  right: 0;
-}
-
-@media (max-width: 1100px) {
-  .logo {
-    display: none;
-  }
+  filter: invert(100%) sepia(0%) saturate(0%) hue-rotate(0deg) brightness(100%) contrast(100%);
+  /* Removed absolute positioning, handled by nav flexbox */
 }
 
 .main {
@@ -576,3 +569,52 @@ pre {
     border-radius: 0.5rem;
     box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 }
+
+.checkboxGroup {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  margin-bottom: 1rem;
+}
+
+.checkboxGroup label {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.errorMessage {
+  color: #e53e3e;
+  font-size: 0.875rem;
+}
+
+/* --- Nav --- */
+.nav {
+  position: relative; /* Context for absolute logo if needed later */
+  display: flex;
+  align-items: center; /* Vertically center items */
+  justify-content: flex-end; /* Pushes items to the right */
+  gap: 16px; /* Space between nav-links div and logo */
+  padding: 16px; /* Match logo margin */
+  height: 64px; /* Logo height (32) + top/bottom padding (16*2) */
+  box-sizing: border-box;
+  width: 100%;
+  background-color: #000; /* Black background */
+  border-bottom: 1px solid #eee; /* Optional: separator */
+}
+
+.nav-links {
+    display: flex;
+    gap: 16px; /* Space between links */
+}
+
+.nav a {
+  color: white;
+  font-weight: bold;
+  text-decoration: none;
+}
+
+.nav a:hover {
+  text-decoration: underline;
+}
+/* --- End Nav --- */
